updating and testin

+++ teaching/vertex_projection_demo.js
@@ -514,7 +514,7 @@ function drawScene(gl, programInfo, buffers, buffers_plane) {
         normalMatrix);
   
     if (document.getElementById("Red").value == 1.0){
-      const temp = parseFloat(document.getElementById("x"))
+      const temp = parseInt(document.getElementById("x"))
       const fieldOfView = 60 * Math.PI / 180;   // in radians
       const aspect = gl.canvas.clientWidth / gl.canvas.clientHeight;
       const zNear = parseFloat(document.getElementById("y").value);
